RiskRadar-Dashboard/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.1
schedule==1.2.0

